[REFACTOR] 패키지 위치에 따른 리팩터링
